Use deviceName for qrlogin

+++ wiliwili/include/activity/live_player_activity.hpp
@@ -119,7 +119,7 @@ private:
     std::map<std::string, LiveDanmakuItemView*> pinnedSuperChatViews;
     
     // 侧边栏弹幕最大数量
-    int maxSidebarDanmakuCount = 100;
+    size_t maxSidebarDanmakuCount = 100;
 
     //更新timeLabel
     MPVEvent::Subscription tl_event_id;
